docs: add MinuteSelectionStep example to MD2/3 demo apps and consolidate viewmodels
